# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ODIN Swagger UI",
        "title": "Odin: Remote Administration for 1S.Enterprise Application Servers",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "khorevaa@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/app/{app}/block": {
            "get": {
                "description": "Установка блокировки на произвольную информационную базу на сервер 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocker"
                ],
                "summary": "Установка блокировки на произвольную информационную базу на сервер 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "infobase uuid or name",
                        "name": "infobase-id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "session deny",
                        "name": "sessions-deny",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "message to user",
                        "name": "message",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "permission code",
                        "name": "permission-code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "denied parameter",
                        "name": "denied-parameter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "permission code",
                        "name": "permission-code",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "scheduled jobs deny",
                        "name": "scheduled-jobs-deny",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"now\"",
                        "description": "denied from time",
                        "name": "denied-from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "denied to time",
                        "name": "denied-to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase user",
                        "name": "infobase-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase password",
                        "name": "infobase-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.InfobaseUnblocker"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Установка блокировки на произвольную информационную базу на сервер 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocker"
                ],
                "summary": "Установка блокировки на произвольную информационную базу на сервер 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "block info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InfobaseBlocker"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cluster uuid",
                        "name": "cluster-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster user",
                        "name": "cluster-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cluster password",
                        "name": "cluster-pwd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase user",
                        "name": "infobase-usr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "infobase password",
                        "name": "infobase-pwd",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.InfobaseUnblocker"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.InfobaseBlocker": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string",
                    "example": "80f7f2f6-2feb-46bf-92f4-19294a2f5dc7"
                },
                "denied_from": {
                    "type": "string",
                    "example": "2020-10-01T08:30:00Z"
                },
                "denied_parameter": {
                    "type": "string",
                    "example": "code"
                },
                "denied_to": {
                    "type": "string",
                    "example": "2020-10-01T08:30:00Z"
                },
                "infobase": {
                    "type": "string",
                    "example": "testdb2"
                },
                "message": {
                    "type": "string",
                    "example": "Обновление информационной базы"
                },
                "permission_code": {
                    "type": "string",
                    "example": "123"
                },
                "reload": {
                    "type": "boolean"
                },
                "scheduled_jobs_deny": {
                    "type": "boolean"
                },
                "sessions_deny": {
                    "type": "boolean"
                }
            }
        },
        "models.InfobaseUnblocker": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string",
                    "example": "80f7f2f6-2feb-46bf-92f4-19294a2f5dc7"
                },
                "denied_parameter": {
                    "type": "string"
                },
                "infobase": {
                    "type": "string",
                    "example": "testdb2"
                },
                "infobase_id": {
                    "type": "string",
                    "example": "80f7f2f6-2feb-46bf-92f4-19294a2f5dc7"
                },
                "permission_code": {
                    "type": "string"
                },
                "scheduled_jobs_deny": {
                    "type": "boolean"
                },
                "sessions_deny": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "metrics": " Grants read access and control metrics operations",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    }
}